django_service/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
pyjwt
gunicorn
python-dotenv
bcrypt
